# docs/schemas/mobile-command.openapi.yaml
openapi: 3.1.0
info:
  title: Lily Mobile Command Pro HTTP API
  version: 1.0.0
  description: Additive v1 contract. Unknown optional fields live only in explicit extensions; unknown mandatory semantics require client upgrade.
servers:
  - url: /public
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, pattern: '^idem_[A-Za-z0-9_-]{8,128}$' } }
    DeviceTimestamp: { name: X-Device-Timestamp, in: header, required: true, schema: { type: string, format: date-time } }
    DeviceNonce: { name: X-Device-Nonce, in: header, required: true, schema: { type: string, minLength: 16, maxLength: 128 } }
    DeviceSignature: { name: X-Device-Signature, in: header, required: true, schema: { type: string, minLength: 64, maxLength: 512 } }
    SignatureVersion: { name: X-Signature-Version, in: header, required: true, schema: { const: '1' } }
    ProtocolVersion: { name: X-Protocol-Version, in: header, required: true, schema: { const: '1' } }
    RemoteSessionId: { name: remoteSessionId, in: path, required: true, schema: { $ref: '#/components/schemas/RemoteSessionId' } }
...
    UploadId: { name: uploadId, in: path, required: true, schema: { $ref: '#/components/schemas/UploadId' } }
    ChunkIndex: { name: chunkIndex, in: path, required: true, schema: { type: integer, minimum: 0, maximum: 99999 } }
    ArtifactId: { name: artifactId, in: path, required: true, schema: { $ref: '#/components/schemas/ArtifactId' } }
  responses:
    Error400: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error401: { description: Authentication, signature, clock, or session failure, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error403: { description: Permission denied or device revoked, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error404: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error409: { description: State, replay, or idempotency conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error410: { description: Resource expired or ended, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error413: { description: Upload too large, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error422: { description: Hash mismatch, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error426: { description: Client upgrade required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error429: { description: Rate or quota limit, headers: { Retry-After: { schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Error503: { description: Dependency unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  schemas:
...
    UnixMs: { type: integer, minimum: 0, maximum: 9007199254740991 }
    Sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    DesktopDeviceId: { type: string, pattern: '^desktop_[A-Za-z0-9_-]+$' }
    MobileDeviceId: { type: string, pattern: '^mobile_[A-Za-z0-9_-]+$' }
...
    ConsumePairingRequest: { type: object, required: [schemaVersion, pairingToken, mobileDeviceId, mobileDeviceName, mobilePlatform, mobilePublicKey, appVersion, supportedProtocolVersions], properties: { schemaVersion: { const: 1 }, pairingToken: { type: string, minLength: 43, maxLength: 512 }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, mobileDeviceName: { type: string, minLength: 1, maxLength: 120 }, mobilePlatform: { enum: [ios, android, web] }, mobilePublicKey: { type: string, minLength: 32, maxLength: 512 }, appVersion: { type: string, minLength: 1, maxLength: 64 }, supportedProtocolVersions: { type: array, minItems: 1, maxItems: 8, uniqueItems: true, items: { type: integer, minimum: 1 } }, extensions: { $ref: '#/components/schemas/Extensions' } }, additionalProperties: false }
    PairingPending: { type: object, required: [schemaVersion, pairingGrantId, mobileDeviceId, desktopDeviceId, status, approvalExpiresAt], properties: { schemaVersion: { const: 1 }, pairingGrantId: { type: string, pattern: '^grant_' }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, desktopDeviceId: { $ref: '#/components/schemas/DesktopDeviceId' }, status: { const: pending_desktop_approval }, approvalExpiresAt: { $ref: '#/components/schemas/UnixMs' } }, additionalProperties: false }
...
    CreateRemoteSessionRequest: { type: object, required: [schemaVersion, desktopDeviceId, mobileDeviceId, pairingGrantId, requestedLevel], properties: { schemaVersion: { const: 1 }, desktopDeviceId: { $ref: '#/components/schemas/DesktopDeviceId' }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, pairingGrantId: { type: string, pattern: '^grant_' }, requestedLevel: { const: 1 }, preferredLilySessionId: { type: string, minLength: 1 }, extensions: { $ref: '#/components/schemas/Extensions' } }, additionalProperties: false }
    RemoteSession: { type: object, required: [schemaVersion, remoteSessionId, desktopDeviceId, mobileDeviceId, permissionLevel, accessToken, refreshToken, expiresAt, commandWsUrl], properties: { schemaVersion: { const: 1 }, remoteSessionId: { $ref: '#/components/schemas/RemoteSessionId' }, desktopDeviceId: { $ref: '#/components/schemas/DesktopDeviceId' }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, lilySessionId: { type: string }, permissionLevel: { $ref: '#/components/schemas/PermissionLevel' }, accessToken: { type: string, minLength: 32 }, refreshToken: { type: string, minLength: 43 }, expiresAt: { $ref: '#/components/schemas/UnixMs' }, commandWsUrl: { type: string, format: uri, pattern: '^wss://' } }, additionalProperties: false }
    RefreshRemoteSessionRequest: { type: object, required: [schemaVersion, mobileDeviceId, refreshToken, refreshTokenGeneration], properties: { schemaVersion: { const: 1 }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, refreshToken: { type: string, minLength: 43 }, refreshTokenGeneration: { type: integer, minimum: 1 } }, additionalProperties: false }
    AccessToken: { type: object, required: [schemaVersion, accessToken, refreshToken, refreshTokenGeneration, expiresAt], properties: { schemaVersion: { const: 1 }, accessToken: { type: string, minLength: 32 }, refreshToken: { type: string, minLength: 43 }, refreshTokenGeneration: { type: integer, minimum: 2 }, expiresAt: { $ref: '#/components/schemas/UnixMs' } }, additionalProperties: false }
    EndRemoteSessionRequest: { type: object, required: [schemaVersion, reason], properties: { schemaVersion: { const: 1 }, reason: { enum: [user_requested, desktop_requested, logout, security, replaced] } }, additionalProperties: false }
    EndReceipt: { type: object, required: [schemaVersion, remoteSessionId, status, endedAt], properties: { schemaVersion: { const: 1 }, remoteSessionId: { $ref: '#/components/schemas/RemoteSessionId' }, status: { const: ended }, endedAt: { $ref: '#/components/schemas/UnixMs' } }, additionalProperties: false }
    PermissionRequest: { type: object, required: [schemaVersion, requestedLevel, reason, ttlSeconds], properties: { schemaVersion: { const: 1 }, requestedLevel: { $ref: '#/components/schemas/PermissionLevel' }, reason: { type: string, minLength: 1, maxLength: 500 }, ttlSeconds: { type: integer, minimum: 1, maximum: 3600 }, screenSourceId: { type: string, minLength: 1, maxLength: 256 } }, additionalProperties: false }
    PermissionProjection:
      oneOf:
        - { $ref: '#/components/schemas/PermissionGranted' }
        - { $ref: '#/components/schemas/PermissionPending' }
        - { $ref: '#/components/schemas/PermissionDenied' }
        - { $ref: '#/components/schemas/PermissionExpired' }
        - { $ref: '#/components/schemas/PermissionRevoked' }
...
    TurnCredentialRequest: { type: object, required: [schemaVersion, peerRole], properties: { schemaVersion: { const: 1 }, peerRole: { enum: [mobile, desktop] } }, additionalProperties: false }
    TurnCredentialSet: { type: object, required: [schemaVersion, iceServers, expiresAt], properties: { schemaVersion: { const: 1 }, iceServers: { type: array, minItems: 1, maxItems: 8, items: { type: object, required: [urls], properties: { urls: { type: array, minItems: 1, maxItems: 8, items: { type: string, format: uri } }, username: { type: string, maxLength: 512 }, credential: { type: string, maxLength: 512 } }, additionalProperties: false } }, expiresAt: { $ref: '#/components/schemas/UnixMs' } }, additionalProperties: false }
    CreateUploadRequest: { type: object, required: [schemaVersion, remoteSessionId, mobileDeviceId, originalName, mimeType, sizeBytes, sha256, chunkSize, chunkCount], properties: { schemaVersion: { const: 1 }, remoteSessionId: { $ref: '#/components/schemas/RemoteSessionId' }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, lilySessionId: { type: string }, originalName: { type: string, minLength: 1, maxLength: 255 }, mimeType: { type: string, minLength: 1, maxLength: 255 }, sizeBytes: { type: integer, minimum: 1, maximum: 524288000 }, sha256: { $ref: '#/components/schemas/Sha256' }, chunkSize: { type: integer, minimum: 1, maximum: 26214400 }, chunkCount: { type: integer, minimum: 1, maximum: 100000 }, extensions: { $ref: '#/components/schemas/Extensions' } }, additionalProperties: false }
    UploadChunk: { type: string, format: binary, minLength: 1, maxLength: 26214400 }
    CompleteUploadRequest: { type: object, required: [schemaVersion, sha256, chunkCount], properties: { schemaVersion: { const: 1 }, sha256: { $ref: '#/components/schemas/Sha256' }, chunkCount: { type: integer, minimum: 1, maximum: 100000 } }, additionalProperties: false }
    UploadRecord: { type: object, required: [schemaVersion, uploadId, remoteSessionId, mobileDeviceId, originalName, displayName, mimeType, sizeBytes, sha256, chunkSize, chunkCount, uploadedChunks, status, risk, expiresAt, createdAt, updatedAt], properties: { schemaVersion: { const: 1 }, uploadId: { $ref: '#/components/schemas/UploadId' }, remoteSessionId: { $ref: '#/components/schemas/RemoteSessionId' }, lilySessionId: { type: string }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, originalName: { type: string, maxLength: 255 }, displayName: { type: string, maxLength: 255 }, mimeType: { type: string, maxLength: 255 }, sizeBytes: { type: integer, minimum: 1, maximum: 524288000 }, sha256: { $ref: '#/components/schemas/Sha256' }, chunkSize: { type: integer, minimum: 1, maximum: 26214400 }, chunkCount: { type: integer, minimum: 1 }, uploadedChunks: { type: array, uniqueItems: true, items: { type: integer, minimum: 0 } }, status: { enum: [created, uploading, uploaded, verifying, verified, pulling, staging, staged, attached, recoverable, cancelled, expired, revoked, failed] }, risk: { type: object, required: [level, reasons, requiresApproval], properties: { level: { enum: [low, medium, high, critical] }, reasons: { type: array, maxItems: 32, items: { type: string, maxLength: 128 } }, requiresApproval: { type: boolean } }, additionalProperties: false }, expiresAt: { $ref: '#/components/schemas/UnixMs' }, createdAt: { $ref: '#/components/schemas/UnixMs' }, updatedAt: { $ref: '#/components/schemas/UnixMs' } }, additionalProperties: false }
    ArtifactDownloadRequest: { type: object, required: [schemaVersion, remoteSessionId, mobileDeviceId], properties: { schemaVersion: { const: 1 }, remoteSessionId: { $ref: '#/components/schemas/RemoteSessionId' }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' } }, additionalProperties: false }
    AuthorizedDownload: { type: object, required: [schemaVersion, artifactId, downloadUrl, expiresAt], properties: { schemaVersion: { const: 1 }, artifactId: { $ref: '#/components/schemas/ArtifactId' }, downloadUrl: { type: string, format: uri, pattern: '^https://' }, expiresAt: { $ref: '#/components/schemas/UnixMs' }, headers: { type: object, additionalProperties: { type: string, maxLength: 4096 } } }, additionalProperties: false }
    ArtifactBytes: { type: string, format: binary }
    PushRegistration: { type: object, required: [schemaVersion, mobileDeviceId, platform, pushToken], properties: { schemaVersion: { const: 1 }, mobileDeviceId: { $ref: '#/components/schemas/MobileDeviceId' }, platform: { enum: [ios, android] }, pushToken: { type: string, minLength: 16, maxLength: 4096 }, environment: { enum: [sandbox, production] } }, additionalProperties: false }
    PushReceipt: { type: object, required: [schemaVersion, registered], properties: { schemaVersion: { const: 1 }, registered: { type: boolean } }, additionalProperties: false }
    DiagnosticsRequest: { type: object, required: [schemaVersion, remoteSessionId, consent, categories], properties: { schemaVersion: { const: 1 }, remoteSessionId: { $ref: '#/components/schemas/RemoteSessionId' }, consent: { const: true }, categories: { type: array, minItems: 1, maxItems: 8, uniqueItems: true, items: { enum: [network, signaling, upload, lifecycle, app] } } }, additionalProperties: false }
...
paths:
...
  /mobile/pairing/consume:
    post: { operationId: consumeMobilePairing, parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ConsumePairingRequest' } } } }, responses: { '202': { description: Pending desktop approval, content: { application/json: { schema: { $ref: '#/components/schemas/PairingPending' } } } }, '400': { $ref: '#/components/responses/Error400' }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '409': { $ref: '#/components/responses/Error409' }, '410': { $ref: '#/components/responses/Error410' } } }
...
  /mobile/sessions:
    post: { operationId: createRemoteSession, parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateRemoteSessionRequest' } } } }, responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/RemoteSession' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '409': { $ref: '#/components/responses/Error409' }, '426': { $ref: '#/components/responses/Error426' } } }
  /mobile/sessions/{remoteSessionId}/refresh:
    post: { operationId: refreshRemoteSession, parameters: [{ $ref: '#/components/parameters/RemoteSessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRemoteSessionRequest' } } } }, responses: { '200': { description: Rotated tokens, content: { application/json: { schema: { $ref: '#/components/schemas/AccessToken' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '409': { $ref: '#/components/responses/Error409' }, '410': { $ref: '#/components/responses/Error410' } } }
  /mobile/sessions/{remoteSessionId}:
    delete: { operationId: endRemoteSession, parameters: [{ $ref: '#/components/parameters/RemoteSessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EndRemoteSessionRequest' } } } }, responses: { '200': { description: Ended, content: { application/json: { schema: { $ref: '#/components/schemas/EndReceipt' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '410': { $ref: '#/components/responses/Error410' } } }
  /mobile/sessions/{remoteSessionId}/permissions:
    post: { operationId: requestRemotePermission, parameters: [{ $ref: '#/components/parameters/RemoteSessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PermissionRequest' } } } }, responses: { '200': { description: Permission projection, content: { application/json: { schema: { $ref: '#/components/schemas/PermissionProjection' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '409': { $ref: '#/components/responses/Error409' }, '503': { $ref: '#/components/responses/Error503' } } }
  /mobile/sessions/{remoteSessionId}/turn-credentials:
    post: { operationId: createTurnCredentials, parameters: [{ $ref: '#/components/parameters/RemoteSessionId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TurnCredentialRequest' } } } }, responses: { '200': { description: Ephemeral TURN set, content: { application/json: { schema: { $ref: '#/components/schemas/TurnCredentialSet' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '503': { $ref: '#/components/responses/Error503' } } }
  /mobile/uploads:
    post: { operationId: createMobileUpload, parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateUploadRequest' } } } }, responses: { '201': { description: Upload created, content: { application/json: { schema: { $ref: '#/components/schemas/UploadRecord' } } } }, '401': { $ref: '#/components/responses/Error401' }, '409': { $ref: '#/components/responses/Error409' }, '413': { $ref: '#/components/responses/Error413' }, '429': { $ref: '#/components/responses/Error429' } } }
  /mobile/uploads/{uploadId}/chunks/{chunkIndex}:
    put: { operationId: putMobileUploadChunk, parameters: [{ $ref: '#/components/parameters/UploadId' }, { $ref: '#/components/parameters/ChunkIndex' }, { $ref: '#/components/parameters/IdempotencyKey' }, { name: X-Chunk-Sha256, in: header, required: true, schema: { $ref: '#/components/schemas/Sha256' } }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/octet-stream: { schema: { $ref: '#/components/schemas/UploadChunk' } } } }, responses: { '200': { description: Chunk committed, content: { application/json: { schema: { $ref: '#/components/schemas/UploadRecord' } } } }, '401': { $ref: '#/components/responses/Error401' }, '409': { $ref: '#/components/responses/Error409' }, '410': { $ref: '#/components/responses/Error410' }, '422': { $ref: '#/components/responses/Error422' } } }
  /mobile/uploads/{uploadId}/complete:
    post: { operationId: completeMobileUpload, parameters: [{ $ref: '#/components/parameters/UploadId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CompleteUploadRequest' } } } }, responses: { '200': { description: Verification checkpoint, content: { application/json: { schema: { $ref: '#/components/schemas/UploadRecord' } } } }, '401': { $ref: '#/components/responses/Error401' }, '409': { $ref: '#/components/responses/Error409' }, '410': { $ref: '#/components/responses/Error410' }, '422': { $ref: '#/components/responses/Error422' } } }
  /mobile/uploads/{uploadId}:
    get: { operationId: getMobileUpload, parameters: [{ $ref: '#/components/parameters/UploadId' }], responses: { '200': { description: Upload checkpoint, content: { application/json: { schema: { $ref: '#/components/schemas/UploadRecord' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '404': { $ref: '#/components/responses/Error404' }, '410': { $ref: '#/components/responses/Error410' } } }
  /mobile/artifacts/{artifactId}/download:
    post: { operationId: authorizeArtifactDownload, parameters: [{ $ref: '#/components/parameters/ArtifactId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ArtifactDownloadRequest' } } } }, responses: { '200': { description: Short-lived authorized download, content: { application/json: { schema: { $ref: '#/components/schemas/AuthorizedDownload' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '404': { $ref: '#/components/responses/Error404' }, '410': { $ref: '#/components/responses/Error410' } } }
  /mobile/artifacts/{artifactId}/content:
    get:
      operationId: getAuthorizedArtifactContent
      description: Fetch bytes using the short-lived bearer authorization returned by authorizeArtifactDownload; authorization remains bound to account, mobile device, artifact, and expiry.
      security:
        - bearerAuth: []
      parameters:
        - { $ref: '#/components/parameters/ArtifactId' }
        - { $ref: '#/components/parameters/ProtocolVersion' }
      responses:
        '200':
          description: Authorized artifact bytes
          headers:
            Content-Disposition: { schema: { type: string, maxLength: 1024 } }
            Content-Length: { schema: { type: integer, minimum: 0 } }
            ETag: { schema: { type: string, maxLength: 128 } }
            Cache-Control: { schema: { const: 'private, no-store' } }
          content:
            application/octet-stream:
              schema: { $ref: '#/components/schemas/ArtifactBytes' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '404': { $ref: '#/components/responses/Error404' }
        '410': { $ref: '#/components/responses/Error410' }
  /mobile/push-token:
    post: { operationId: registerMobilePushToken, parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PushRegistration' } } } }, responses: { '200': { description: Registered, content: { application/json: { schema: { $ref: '#/components/schemas/PushReceipt' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '409': { $ref: '#/components/responses/Error409' } } }
    delete: { operationId: unregisterMobilePushToken, parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], responses: { '200': { description: Unregistered, content: { application/json: { schema: { $ref: '#/components/schemas/PushReceipt' } } } }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' } } }
  /mobile/diagnostics:
    post: { operationId: createMobileDiagnostics, parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/DeviceTimestamp' }, { $ref: '#/components/parameters/DeviceNonce' }, { $ref: '#/components/parameters/DeviceSignature' }, { $ref: '#/components/parameters/SignatureVersion' }, { $ref: '#/components/parameters/ProtocolVersion' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DiagnosticsRequest' } } } }, responses: { '202': { description: Redacted diagnostic bundle accepted, content: { application/json: { schema: { $ref: '#/components/schemas/DiagnosticsBundle' } } } }, '400': { $ref: '#/components/responses/Error400' }, '401': { $ref: '#/components/responses/Error401' }, '403': { $ref: '#/components/responses/Error403' }, '503': { $ref: '#/components/responses/Error503' } } }